Add simple-require-loader (#881)
It's literally a 528-byte fully-featured module loader (a la AMD or CommonJS), but has a ton of features, including cross-platform, async remote loading support in both the browser environment and workers. Another version (not directly linked here) is 340 bytes, and doesn't include remote loading support (but it has everything else).