Please add FastestSmallestTextEncoderDecoder; BestBase64EncoderDecoder; Fast Map, WeakMap, Set, and WeakSet Polyfill; SuperSimpleExtensibleSmallXMLHttpRequestWrapper; Force DOM reflow JS; DeferStackJS; AsyncGlobalEvalFunction; and SPromiseMeSpeedJS
I have included the minified files instead of the source files because Closure Compiler is so much better than UglifyJS (especially for the way in which I code Javascript) that the representation of the size by UglifyJS would be completely unfair. For example, I use a DEBUGMODE switch variable in SPromiseMeSpeed to easily differentiate and distribute two separate versions of the library. UglifyJS is simply too primitive to properly optimize this Javascript code so completely as Closure Compiler can.  In the "Rules" section, you say that "hand-optimized" code is preferred. Well, I could spent an infinite amount of time hand-optimizing the minified size of my libraries and still not beat Closure Compiler. Even after much scrutiny on every minified file, the only three problem I have yet to find are that `typeof global==="undefined"` is not transformed into `typeof global==void 0`, `Infinity` is not transformed into `1/0`, and `NaN` is not transformed into `0/0`. I submitted these bugs to the [Closure Compiler repository](https://github.com/google/closure-compiler/issues/2934), but they did not seem interested. Thus, I hand-optimize these particular issues instead.

The solution to Closure Compiler with my libraries is to provide the code minified by Closure Compiler to UglifyJS. UglifyJS shouldn't increase the file size (of the files already minified by closure compiler) too much, so using the minified files as the source files ought to be a fair compromise.

Further, another reason for these dropbox links is that they allow me to distribute new versions of the libraries for bug fixes and updating the size here on microjs. This may sound evil, but really I assure you it is not. I have no big plans or intentions for any of these libraries, so any new changes that I would post up would only be bug fixes. I would never post up breaking changes for any of these libraries.

Finally, I petition the inclusion of my libraries into the MicroJS list because of the fact that all of my libraries in this pull request are under 5kb before gzip.