Add tint to the list
`tint` is a template engine. And you are very clear about the templating engines in README.md. I agree with that.

The reason I insist on this pull request is that `tint` is not a normal template engine. It doesn't render the result to a string, it compiles the model to hyperscript functions. This is a very niche usage.

And using hyperscript to create views is quite confusing for designers. This is a topic of much discussion in the community of these frameworks.

There are many frameworks like hyperapp, superfine, mithril, preact, etc that use hyperscript to create their own views.

This is a very beautiful, small and specific software that I believe its dissemination will bring great benefits to the community of these frameworks.